+++ oci-test/tests/oci_cloud_guard_configuration/test-list-query.sql
@@ -1,3 +1,3 @@
-select reporting_region, self_manage_resources
+select reporting_region, status
 from oci.oci_cloud_guard_configuration
-where region = '{{ output.reporting_region.value }}';
+where reporting_region = '{{ output.reporting_region.value }}';