+++ oci-test/tests/oci_core_vnic_attachment/test-turbot-query.sql
@@ -1,0 +1,3 @@
+select title, tenant_id
+from oci_core_vnic_attachment
+where id = '{{ output.resource_id.value }}';