+++ oci-test/tests/oci_vault_secret/test-list-query.sql
@@ -1,0 +1,3 @@
+select name, id, lifecycle_state
+from oci_vault_secret
+where name = '{{ output.resource_name.value }}';