+++ oci-test/tests/oci_vault_secret/test-turbot-query.sql
@@ -1,0 +1,3 @@
+select title, tenant_id
+from oci_vault_secret
+where id = '{{ output.resource_id.value }}';