+++ oci/table_oci_cloud_guard_managed_list.go
@@ -7,9 +7,9 @@ import (
 	"github.com/oracle/oci-go-sdk/v44/cloudguard"
 	"github.com/oracle/oci-go-sdk/v44/common"
 	"github.com/turbot/go-kit/types"
-	"github.com/turbot/steampipe-plugin-sdk/grpc/proto"
-	"github.com/turbot/steampipe-plugin-sdk/plugin"
-	"github.com/turbot/steampipe-plugin-sdk/plugin/transform"
+	"github.com/turbot/steampipe-plugin-sdk/v2/grpc/proto"
+	"github.com/turbot/steampipe-plugin-sdk/v2/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/v2/plugin/transform"
 )
 
 //// TABLE DEFINITION
@@ -166,7 +166,7 @@ func tableCloudGuardManagedList(_ context.Context) *plugin.Table {
 
 //// LIST FUNCTION
 
-func listCloudGuardManagedLists(ctx context.Context, d *plugin.QueryData, _ *plugin.HydrateData) (interface{}, error) {
+func listCloudGuardManagedLists(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
 	logger := plugin.Logger(ctx)
 	compartment := plugin.GetMatrixItem(ctx)[matrixKeyCompartment].(string)
 	logger.Debug("oci.listCloudGuardManagedLists", "Compartment", compartment)
@@ -178,12 +178,21 @@ func listCloudGuardManagedLists(ctx context.Context, d *plugin.QueryData, _ *plu
 		return nil, nil
 	}
 
-	// Create Session
-	session, err := cloudGuardService(ctx, d)
+	// fetch reporting region from configuration
+	getCloudGuardConfigurationCached := plugin.HydrateFunc(getCloudGuardConfiguration).WithCache()
+	configuration, err := getCloudGuardConfigurationCached(ctx, d, h)
 	if err != nil {
 		return nil, err
 	}
-	
+
+	reportingRegion := configuration.(cloudguard.Configuration).ReportingRegion
+
+	// Create Session
+	session, err := cloudGuardService(ctx, d, *reportingRegion)
+	if err != nil {
+		return nil, err
+	}
+
 	// Build request parameters
 	request := buildCloudGuardManagedListFilters(equalQuals)
 	request.CompartmentId = types.String(compartment)
@@ -241,8 +250,18 @@ func getCloudGuardManagedList(ctx context.Context, d *plugin.QueryData, h *plugi
 	if id == "" {
 		return nil, nil
 	}
+
+	// fetch reporting region from configuration
+	getCloudGuardConfigurationCached := plugin.HydrateFunc(getCloudGuardConfiguration).WithCache()
+	configuration, err := getCloudGuardConfigurationCached(ctx, d, h)
+	if err != nil {
+		return nil, err
+	}
+
+	reportingRegion := configuration.(cloudguard.Configuration).ReportingRegion
+
 	// Create Session
-	session, err := cloudGuardService(ctx, d)
+	session, err := cloudGuardService(ctx, d, *reportingRegion)
 	if err != nil {
 		return nil, err
 	}