+++ oci/table_oci_mysql_db_system_metric_connections_daily.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 
 	"github.com/oracle/oci-go-sdk/v44/mysql"
-	"github.com/turbot/steampipe-plugin-sdk/grpc/proto"
-	"github.com/turbot/steampipe-plugin-sdk/plugin"
-	"github.com/turbot/steampipe-plugin-sdk/plugin/transform"
+	"github.com/turbot/steampipe-plugin-sdk/v2/grpc/proto"
+	"github.com/turbot/steampipe-plugin-sdk/v2/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/v2/plugin/transform"
 )
 
 //// TABLE DEFINITION
@@ -42,7 +42,7 @@ func listMySQLDBSystemMetricConnectionsDaily(ctx context.Context, d *plugin.Quer
 	}
 
 	_, err := listMonitoringMetricStatistics(ctx, d, "Daily", "oci_mysql_database", "ActiveConnections", "resourceId", *dbSystem.Id, *dbSystem.CompartmentId, region)
-	if  err != nil {
+	if err != nil {
 		return nil, err
 	}