+++ oci/table_oci_mysql_db_system_metric_cpu_utilization_hourly.go
@@ -1,0 +1,45 @@
+package oci
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/oracle/oci-go-sdk/v44/mysql"
+	"github.com/turbot/steampipe-plugin-sdk/v2/grpc/proto"
+	"github.com/turbot/steampipe-plugin-sdk/v2/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/v2/plugin/transform"
+)
+
+//// TABLE DEFINITION
+
+func tableOciMySQLDBSystemMetricCpuUtilizationHourly(_ context.Context) *plugin.Table {
+	return &plugin.Table{
+		Name:        "oci_mysql_db_system_metric_cpu_utilization_hourly",
+		Description: "OCI MySQL DB System Monitoring Metrics - CPU Utilization (Hourly)",
+		List: &plugin.ListConfig{
+			ParentHydrate: listMySQLDBSystems,
+			Hydrate:       listMySQLDBSystemMetricCpuUtilizationHourly,
+		},
+		GetMatrixItem: BuildCompartementRegionList,
+		Columns: MonitoringMetricColumns(
+			[]*plugin.Column{
+				{
+					Name:        "id",
+					Description: "The OCID of the DB System.",
+					Type:        proto.ColumnType_STRING,
+					Transform:   transform.FromField("DimensionValue"),
+				},
+			}),
+	}
+}
+
+func listMySQLDBSystemMetricCpuUtilizationHourly(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
+	dbSystem := h.Item.(mysql.DbSystemSummary)
+	region := fmt.Sprintf("%v", ociRegionNameFromId(*dbSystem.Id))
+
+	if dbSystem.LifecycleState == "DELETING" || dbSystem.LifecycleState == "DELETED" {
+		return nil, nil
+	}
+
+	return listMonitoringMetricStatistics(ctx, d, "HOURLY", "oci_mysql_database", "CPUUtilization", "resourceId", *dbSystem.Id, *dbSystem.CompartmentId, region)
+}